# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Movie Search API",
    "version": "1.0.0",
    "description": "Public REST API for semantic movie search. Wraps FastMCP movie-search tools behind JWT Bearer auth (client credentials). Minimal APIs on .NET 10."
  },
  "servers": [
    { "url": "http://localhost:8080", "description": "Local Docker Compose" }
  ],
  "tags": [
    { "name": "Auth" },
    { "name": "Health" },
    { "name": "Movies" }
  ],
  "components": {
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT from POST /auth/token"
      }
    },
    "schemas": {
      "Movie": {
        "type": "object",
        "required": ["id", "title"],
        "properties": {
          "id": { "type": "string", "examples": ["11111111-1111-1111-1111-111111111111"] },
          "title": { "type": "string", "examples": ["The Matrix"] },
          "releaseYear": { "type": ["integer", "null"], "examples": [1999] },
          "majorGenre": { "type": ["string", "null"], "examples": ["Action"] },
          "mpaaRating": { "type": ["string", "null"], "examples": ["R"] },
          "director": { "type": ["string", "null"], "examples": ["Lana Wachowski"] },
          "distributor": { "type": ["string", "null"], "examples": ["Warner Bros."] },
          "imdbRating": { "type": ["number", "null"], "examples": [8.7] },
          "rtRating": { "type": ["integer", "null"], "examples": [87] },
          "similarity": { "type": ["number", "null"], "description": "Cosine similarity score", "examples": [0.91] }
        }
      },
      "DatasetStats": {
        "type": "object",
        "required": ["totalMovies", "genres"],
        "properties": {
          "totalMovies": { "type": "integer", "examples": [3201] },
          "genres": { "type": "integer", "examples": [12] },
          "yearMin": { "type": ["integer", "null"], "examples": [1920] },
          "yearMax": { "type": ["integer", "null"], "examples": [2011] },
          "avgImdbRating": { "type": ["number", "null"], "examples": [6.4] }
        }
      },
      "TokenRequest": {
        "type": "object",
        "required": ["grant_type", "client_id", "client_secret"],
        "properties": {
          "grant_type": { "type": "string", "examples": ["client_credentials"] },
          "client_id": { "type": "string", "examples": ["reader"] },
          "client_secret": { "type": "string", "examples": ["reader-secret"] }
        }
      },
      "TokenResponse": {
        "type": "object",
        "required": ["access_token", "token_type", "expires_in", "role"],
        "properties": {
          "access_token": { "type": "string" },
          "token_type": { "type": "string", "examples": ["Bearer"] },
          "expires_in": { "type": "integer", "examples": [3600] },
          "role": { "type": "string", "enum": ["reader", "admin"] }
        }
      },
      "Problem": {
        "type": "object",
        "properties": {
          "type": { "type": "string" },
          "title": { "type": "string" },
          "status": { "type": "integer" },
          "detail": { "type": "string" }
        }
      }
    }
  },
  "paths": {
    "/health": {
      "get": {
        "tags": ["Health"],
        "summary": "Liveness probe",
        "operationId": "HealthLive",
        "responses": {
          "200": {
            "description": "Process is up",
            "content": {
              "application/json": {
                "example": { "status": "healthy", "checks": { "mcp": "fake" } }
              }
            }
          }
        }
      }
    },
    "/health/ready": {
      "get": {
        "tags": ["Health"],
        "summary": "Readiness probe",
        "operationId": "HealthReady",
        "responses": {
          "200": { "description": "Ready to serve traffic" },
          "503": { "description": "MCP client configured and unreachable" }
        }
      }
    },
    "/auth/token": {
      "post": {
        "tags": ["Auth"],
        "summary": "Issue a JWT via client credentials",
        "operationId": "IssueToken",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/TokenRequest" },
              "example": {
                "grant_type": "client_credentials",
                "client_id": "reader",
                "client_secret": "reader-secret"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Token issued",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/TokenResponse" } }
            }
          },
          "400": { "description": "Malformed request", "content": { "application/problem+json": { "schema": { "$ref": "#/components/schemas/Problem" } } } },
          "401": { "description": "Invalid credentials", "content": { "application/problem+json": { "schema": { "$ref": "#/components/schemas/Problem" } } } }
        }
      }
    },
    "/api/v1/movies/search": {
      "get": {
        "tags": ["Movies"],
        "summary": "Natural-language movie search",
        "description": "Reader or admin. Maps to MCP search_movies_by_description.",
        "operationId": "SearchMovies",
        "security": [{ "Bearer": [] }],
        "parameters": [
          { "name": "q", "in": "query", "required": true, "schema": { "type": "string" }, "example": "action movies from the 90s with high IMDB ratings" },
          { "name": "top_k", "in": "query", "schema": { "type": "integer", "default": 10, "minimum": 1, "maximum": 50 } },
          { "name": "genre", "in": "query", "schema": { "type": "string" }, "example": "Action" },
          { "name": "min_imdb_rating", "in": "query", "schema": { "type": "number" }, "example": 7.5 },
          { "name": "mpaa_rating", "in": "query", "schema": { "type": "string" }, "example": "R" },
          { "name": "decade", "in": "query", "schema": { "type": "integer" }, "example": 1990 }
        ],
        "responses": {
          "200": {
            "description": "Ranked results",
            "content": {
              "application/json": {
                "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Movie" } },
                "example": [
                  {
                    "id": "11111111-1111-1111-1111-111111111111",
                    "title": "The Matrix",
                    "releaseYear": 1999,
                    "majorGenre": "Action",
                    "mpaaRating": "R",
                    "director": "Lana Wachowski",
                    "distributor": "Warner Bros.",
                    "imdbRating": 8.7,
                    "rtRating": 87,
                    "similarity": 0.91
                  }
                ]
              }
            }
          },
          "400": { "description": "Missing q" },
          "401": { "description": "Missing or invalid token" },
          "429": { "description": "Rate limit exceeded" }
        }
      }
    },
    "/api/v1/movies/{id}": {
      "get": {
        "tags": ["Movies"],
        "summary": "Get a movie by ID",
        "description": "Admin only. Maps to MCP get_movie_by_id (Part 3 contract addition).",
        "operationId": "GetMovieById",
        "security": [{ "Bearer": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Movie", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Movie" } } } },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Reader tokens are not permitted" },
          "404": { "description": "Unknown id" }
        }
      }
    },
    "/api/v1/movies/{id}/similar": {
      "get": {
        "tags": ["Movies"],
        "summary": "Semantically similar movies",
        "description": "Admin only. Maps to MCP get_similar_movies.",
        "operationId": "GetSimilarMovies",
        "security": [{ "Bearer": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "top_k", "in": "query", "schema": { "type": "integer", "default": 5, "maximum": 50 } }
        ],
        "responses": {
          "200": { "description": "Similar movies", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Movie" } } } } },
          "403": { "description": "Reader tokens are not permitted" }
        }
      }
    },
    "/api/v1/movies/genres": {
      "get": {
        "tags": ["Movies"],
        "summary": "List distinct genres",
        "description": "Admin only. Maps to MCP list_genres.",
        "operationId": "ListGenres",
        "security": [{ "Bearer": [] }],
        "responses": {
          "200": { "description": "Genre names", "content": { "application/json": { "schema": { "type": "array", "items": { "type": "string" } }, "example": ["Action", "Drama"] } } },
          "403": { "description": "Reader tokens are not permitted" }
        }
      }
    },
    "/api/v1/stats": {
      "get": {
        "tags": ["Movies"],
        "summary": "Dataset statistics",
        "description": "Admin only. Maps to MCP get_dataset_stats.",
        "operationId": "GetDatasetStats",
        "security": [{ "Bearer": [] }],
        "responses": {
          "200": { "description": "Stats", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/DatasetStats" } } } },
          "403": { "description": "Reader tokens are not permitted" }
        }
      }
    }
  }
}